feat(sourcerer): IdentityMediator API simplification
Introduce getClaims, authorize and revoke
Add authorizeURI for identity.js

+++ packages/gateway-api/identity.js
@@ -2,21 +2,27 @@ const { base64URLDecode } = require("./base64");
 const inBrowser = require("./inBrowser").default;
 
 module.exports = {
+  getClientId,
   getClientCookieName,
   getSessionCookieName,
   getSessionClaims,
   authorizeSession,
+  revokeSession,
   finalizeSession,
 };
 
+function getClientId (options = {}) {
+  return options.clientId || options.clientURI || (this || {}).clientId || (this || {}).clientURI;
+}
+
 function getClientCookieName (options = {}) {
-  if (!options.client_id) options.client_id = options.clientURI || (this || {}).clientURI;
+  if (!options.client_id) options.client_id = getClientId.call(this, options);
   if (!options.client_id) throw new Error("getClientCookieName.(clientURI|client_id) missing");
   return `__Secure-valos-client-${encodeURIComponent(options.client_id)}`;
 }
 
 function getSessionCookieName (options = {}) {
-  if (!options.client_id) options.client_id = options.clientURI || (this || {}).clientURI;
+  if (!options.client_id) options.client_id = getClientId.call(this, options);
   if (!options.client_id) throw new Error("getSessionCookieName.(clientURI|client_id) missing");
   return `__Secure-valos-session-token-${encodeURIComponent(options.client_id)}`;
 }
@@ -36,15 +42,16 @@ function getSessionClaims (options = {}) {
 
 /* eslint-disable camelcase */
 function authorizeSession ({
-  grantProvider, client_id, clientURI, redirect_uri, sessionURI, ...rest
+  grantProvider, client_id, clientId, clientURI, redirect_uri, sessionURI, ...rest
 }) {
   if (!inBrowser()) throw new Error("Cannot authorize a session in a non-browser context");
+  const authorizeURI = grantProvider || (this || {}).authorizeURI;
   if (!grantProvider) throw new Error("authorizeSession.grantProvider missing");
-  if (grantProvider.substr(0, 8) !== "https://") {
+  if (authorizeURI.substr(0, 8) !== "https://") {
     throw new Error("Invalid grant provider: only OpenId Connect https:// scheme is supported");
   }
   const params = {
-    client_id: client_id || clientURI || (this || {}).clientURI,
+    client_id: getClientId.call(this, { clientId: client_id || clientId, clientURI }),
     redirect_uri: redirect_uri || sessionURI || (this || {}).sessionURI,
     ...rest,
   };
@@ -62,7 +69,7 @@ function authorizeSession ({
     window.document.cookie = `${clientCookieName}=${encodeURIComponent(params.state)
       }; max-age=900; Secure; SameSite=Lax`;
   }
-  let requestAuthorization = `${grantProvider}?response_type=code`;
+  let requestAuthorization = `${authorizeURI}?response_type=code`;
   Object.keys(params).forEach(k => {
     if (params[k]) {
       requestAuthorization += `&${encodeURIComponent(k)}=${encodeURIComponent(params[k])}`;
@@ -72,12 +79,19 @@ function authorizeSession ({
 }
 
 function finalizeSession (options = {}) {
-  if (!inBrowser()) throw new Error("Cannot finalize a session in a non-browser context");
-  const sessionURI = options.redirect_uri || options.sessionURI || (this || {}).sessionURI;
-  if (!sessionURI) throw new Error("finalizeSession.sessionURI missing");
-  const ret = window.fetch(sessionURI, {
-    method: "DELETE", credentials: "same-origin", mode: "same-origin", redirect: "error",
-  });
+  return revokeSession.call(this, options);
+}
+
+function revokeSession (options = {}) {
+  let ret;
+  if (options.redirect !== false) {
+    if (!inBrowser()) throw new Error("Cannot finalize a session in a non-browser context");
+    const sessionURI = options.redirect_uri || options.sessionURI || (this || {}).sessionURI;
+    if (!sessionURI) throw new Error("finalizeSession.sessionURI missing");
+    ret = (options.redirect !== false) && window.fetch(sessionURI, {
+      method: "DELETE", credentials: "same-origin", mode: "same-origin", redirect: "error",
+    });
+  }
   window.document.cookie = `${getClientCookieName.call(this, options)}=;max-age=0`;
   window.document.cookie = `${getSessionCookieName.call(this, options)}=;max-age=0`;
   return ret;
